agents/data-scraper/requirements/requirements.txt
# Core dependencies for Data Scraper Agent

# HTTP and async support
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Data validation and settings
pydantic==2.11.7
pydantic-settings==2.10.1

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# HTTP and networking
requests==2.31.0
urllib3==2.0.7

# Date and time handling
python-dateutil==2.8.2

# Retry and resilience
tenacity==8.2.3

# Type hints support
typing-extensions==4.12.2

# Optional: For advanced JSON path processing
jsonpath-ng==1.6.0

# Optional: For XML processing
lxml==4.9.4
